requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-ckeditor==5.9.0
django-filter==2.3.0
django-google-maps==0.12.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-location-field==2.1.0
django-modelcluster==5.0.2
django-multiselectfield==0.1.12
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
easy-thumbnails==2.7
html5lib==1.0.1
idna==2.9
l18n==2018.5
Pillow==7.1.2
psycopg2==2.8.5
python-slugify==4.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.3
XlsxWriter==1.2.9
gunicorn==20.0.4